Moved the agent code into its own class.
