+++ cartpole-v0/main.py
@@ -7,60 +7,9 @@ from keras.layers import Dense
 from keras.optimizers import Adam
 from keras import backend as K
 
+from agent import DQNAgent
+
 EPISODES = 1000
-
-# Deep Q-learning Agent
-class DQNAgent:
-    def __init__(self, state_size, action_size):
-        self.state_size = state_size
-        self.action_size = action_size
-        self.memory = deque(maxlen=2000)
-        self.gamma = 0.95    # discount rate
-        self.epsilon = 1.0  # exploration rate
-        self.epsilon_min = 0.01
-        self.epsilon_decay = 0.995
-        self.learning_rate = 0.001
-        self.model = self._build_model()
-
-    def _build_model(self):
-        # TODO: Rebuild as a Tensorflow model, will probably want to encapsulate a it's own class
-        # Neural Net for Deep-Q learning Model
-        model = Sequential()
-        model.add(Dense(24, input_dim=self.state_size, activation='relu'))
-        model.add(Dense(24, activation='relu'))
-        model.add(Dense(self.action_size, activation='linear'))
-        model.compile(loss='mse',
-                      optimizer=Adam(lr=self.learning_rate))
-        return model
-
-    def remember(self, state, action, reward, next_state, done):
-        self.memory.append((state, action, reward, next_state, done))
-
-    def act(self, state, force_exploitation=False):
-        if np.random.rand() <= self.epsilon and not force_exploitation:
-            return random.randrange(self.action_size)
-
-        act_values = self.model.predict(state)
-
-        return np.argmax(act_values[0])  # returns action
-
-    def replay(self, batch_size):
-        batch_size = min(batch_size, len(self.memory))
-        minibatch = random.sample(self.memory, batch_size)
-
-        for state, action, reward, next_state, done in minibatch:
-            target = reward
-
-            if not done:
-              target = reward + self.gamma * np.amax(self.model.predict(next_state)[0])
-
-            target_f = self.model.predict(state)
-            target_f[0][action] = target
-
-            self.model.fit(state, target_f, epochs=1, verbose=0)
-
-        if self.epsilon > self.epsilon_min:
-            self.epsilon *= self.epsilon_decay
 
 if __name__ == "__main__":
     # initialize gym environment and the agent
